Add .rs source code file of A : B Problem

+++ introduction/ITP1_4_A-A : B Problem.rs
@@ -4,5 +4,5 @@ fn main() {
   io::stdin().read_line(&mut input).unwrap();
   let l:Vec<_> = input.split_whitespace().map(|s| s.parse::<i32>().unwrap()).collect();
   let (a, b) = (l[0], l[1]);
-  println!("{} {} {}", a/b, a%b, a/b);
+  println!("{} {} {}", a/b, a%b, a as f64 / b as f64);
 }
